ironic: always enable conductor HTTP server
In the Xena release, Ironic removed the iSCSI driver [1]. The
recommended driver is direct, which uses HTTP to transfer the disk
image. This requires an HTTP server, and the simplest option is to use
the one currently deployed when enable_ironic_ipxe is set to true. For
this reason, this patch always enables the HTTP server running on the
conductor.

iPXE is still enabled separately, since it cannot currently be used at
the same time as PXE.

Backport note: Ironic made direct the default in Wallaby. This Wallaby
backport ensures that the defaults in Kolla & Ironic may work together.

[1] https://review.opendev.org/c/openstack/ironic/+/789382

Change-Id: I30c2ad2bf2957ac544942aefae8898cdc8a61ec6
(cherry picked from commit 411668ea5a05e8182d80be63501e2f7fa1882345)