xena and earlier: Fix bootstrap servers by installing apparmor package
Probably because of a new version of docker-ce package
apparmor is not installed by default anymore and
containers can't start because of apparmor_parser is missing.

This patch is adding apparmor package to list of packages to
be installed.

This code has been moved to ansible-collection-kolla in Yoga

Closes-Bug: #2004583
Change-Id: Ifcb06f82ce2495ad2d342416865f97b9f94a724a
(cherry picked from commit 80dd08f28710b18778a21c3a602377be46c4c6cf)